Remove address validation in addressController
The address validation function was removed from the 'addAddressHandler' and 'updateAddressHandler' functions and is no longer being used in the controller. It's unclear if this validation moved elsewhere or if it was just deemed unnecessary. With this change, future errors in address input will not be caught in the server controller.